Add figwheel, make sure Austin/CIDER still work.
I am quite pleased with myself :)

As before: M-x cider-jack-in ; (run) ; (browser-repl)
Separate: lein figwheel

The figwheel process will take care of rebuilding the cljs, and it will
start a server that the figwheel client running in the browser can
connect to (port 3449, which is the default, although we've configured
it explicitly for explicitness).

The jetty server from before will serve the app, making sure it connects
to the repl server by injecting the right magic invocation in the index.html.