Make development mode work again, load react from jar
To run with figwheel, we need to disable cljs optimizations. This also
means that we manually need to require our main namespace. Before we had
that code in index.html, but in production we only need to load app.js.

Since we already inject JS for the browser repl, we can expand that to
also inject the necessary includes. I've also added a compojure route to
serve up react from the com.facebook/react gem, so we can serve it
ourselves instead of relying on facebook's CDN.