Switch to Weasel, and other improvements
Switched from Austin to Weasel for a browser-connected REPL. I regularly
got issues with Austin that after a while the REPL became
unresponsive. Hopefully this fixes that. Weasel seems to bring some
other niceties as well.

Did some testing with a project name containing a dash and found some
places where {{name}} had to be {{sanitized}} or vice versa. Switched
the uberjar name to not be sanitized, this is more in lein with the jar
naming leiningen uses by default.

Finally the "is-dev" thing on the clojurescript side now actually
works. The idea is that in dev mode we set an "is-dev" class on the
<body>, so the clojurescript side can know we're runing in dev mode and
needs to connect to the REPL and Figwheel. So in production no
REPL/figwheel connections will be started.