Drop dev/prod directories for a single dev dir
Before we had development helper functions like (run) and (browser-repl)
inside the {{project-ns}}.server namespace, and we would start the repl
with that as the default namespace, so they were available.

In prod mode we don't want to include figwheel though, and there are
potentially other small differences, so we had a separate source
directory for dev and prod, so you could implement a namespace twice,
once for each env.

The newer setup is simpler, we add our convenience functions to `user`,
and start the repl in `user` the way it does by default. The good thing
about this is you have the default clojure repl functions like doc and
source available as well. This dramatically simplifies the directory layout.