Use Figwheel to handle HTTP requests again
By switching to Component we lost something we had previously: Figwheel as the
one and only HTTP server.

Instead Figwheel and jetty/http-kit would both get started. This is confusing
because now in dev mode API calls need to go to a different port. This is also
problematic because when used from the frontend it violates browsers Same Origin
Policy, meaning the frontend can no longer (without extra hassle, i.e. CORS
headers) make backend calls.

The new approach is to have a helper function in `(ns user)`,
`dev-ring-handler`, which takes a request, and delegates it to whatever the
current handler in the system map is. In `project.clj` Figwheel is told to use
this function as its ring handler.

Another change is to use a regular function for the routes factory, rather than
the inline anonymous fn we had earlier. This is more how these Duct-style
endpoint components are intended to be used, and it opens the door to passing
dependencies into the routes.