re-frame support (experimental)
Adds a re-frame flag which adds re-frame as a dependency, and creates the
typical re-frame structure of db/subs/views/events.

This uses the templates directly from re-frame-template. At the moment only the
basic vanilla re-frame template is supported, but it would be easy to add
support for +re-frisk or +re-com.