Fix the pop poll. Use `curl` (yes, really)
Seems Java before 8u101 does not play nicely with Let's Encrypt certs. Tried
hard to make it accept the certificate, couldn't get it to work, so until most
of the world has upgraded (yeah right), just shelling out to curl.

Not everyone has curl either, but that's fine, we just swallow any exceptions.
All we need is a rough indication of which versions and flags are popular.