Compatibility fixes: --reagent and lein figwheel
There's a bunch of existing documentation out there that starts with a `lein new
chestnut`, and so we should try not to break existing invocations and uses. Even
though `--reagent` is now the default, it should still be possible to specify
the flag.

We should also make sure that `lein figwheel` continues to work. Normally we
start figwheel as part of the system, but sometimes people use Chestnut only for
the figwheel/clojurescript part, and in that case they might only care about
running Figwheel. Since we write our compiled js to dev-target/public, we should
make sure dev-target is on the classpath so the Figwheel web server can find it.